tests: Moving the tests to the right package
